Added tests using Test::Kwalitee::Extra which seems to be the distribution currently holding the test which fails. These tests where previously run under TEST_AUTHOR, but that are now enabled using RELEASE_TESTING instead.
The kwalitee tests are only run under this flag and if Test::Kwalitee
and Test::Kwalitee are installed.

I ran into some problems with one of the tests in
Test::Kwalitee::Extra, prereq_matches_use - so this one has been
disabled. It reports that App::Prove is not required but used.

App::Prove is only listed in the Build.PL file, since this is a
Task distribution. Perhaps I am just tired but I will investigate
this further I did not know Test::Kwalitee::Extra before tonight, so
this is my first use of it.